% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refresh.R
\name{data_update}
\alias{data_update}
\title{Update the Package Datasets}
\usage{
data_update()
}
\description{
Checks if new data is available on the package dev version (Github).
In case new data is available the function will enable the user the update the datasets
}
