% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italy_total}
\alias{italy_total}
\title{The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Italy Outbreak Dataset}
\format{
A data.frame object
}
\source{
Presidenza del Consiglio dei Ministri - Dipartimento della Protezione Civile \href{http://www.protezionecivile.it/}{website}
}
\usage{
italy_total
}
\description{
daily summary of the Coronavirus (COVID-19) pandemic cases in Italy
}
\details{
The dataset contains the daily summary of the Coronavirus pandemic cases in Italy. The data includes the following fields:

- date - date of notification (Date class)

- hospitalized_with_symptoms - number of hospitalised patients with symptoms

- intensive_care - intensive Care

- total_hospitalized - total hospitalised patients

- home_confinement - home confinement

- cumulative_positive_cases - total amount of current positive cases (Hospitalised patients + Home confinement)

- daily_positive_cases - new amount of current positive cases (Hospitalised patients + Home confinement)

- recovered - recovered cases

- death - death cases

- positive_clinical_activity - positive cases emerged from clinical activity

- positive_surveys_tests - positive cases emerging from surveys and tests, planned at national or regional level

- cumulative_cases - total positive cases, recovered, and death cases

- total_tests - number of tests performed

- total_people_tested - total number of people tested
}
\examples{

data(italy_total)

head(italy_total)

}
\keyword{COVID19}
\keyword{Italy}
\keyword{coronavirus}
\keyword{datasets}
