% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italy_province}
\alias{italy_province}
\title{The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Italy Provinces Outbreak Dataset}
\format{
A data.frame object
}
\source{
Presidenza del Consiglio dei Ministri - Dipartimento della Protezione Civile \href{http://www.protezionecivile.it/}{website}
}
\usage{
italy_province
}
\description{
daily summary of the Coronavirus (COVID-19) pandemic cases in Italy's provinces
}
\details{
The dataset contains the daily summary of the Coronavirus pandemic cases in Italy's regions. The data includes the following fields:

- date - date of notification (Date class)

- region_code - region code

- region_name - region names

- province_code - province code

- province_name - province names

- lat - latitude point

- long - Longitude point

- total_positive_cases - total positive cases

- total_tests - number of tests performed
}
\examples{

data(italy_province)

head(italy_province)

}
\keyword{COVID19}
\keyword{Italy}
\keyword{coronavirus}
\keyword{datasets}
